[update] JGAMS module
