+++ modules/image_functions.py
@@ -125,24 +125,6 @@ class ImageFunctions(DNNFunctions):
 
         
         self.sam_mode = False
-
-        """
-        Pompt Tool
-        """
-        mainWidgets.brushButton.clicked.connect(self.openBrushMenu)
-
-        self.BrushMenu = DinoPromptWindow()
-        self.BrushMenu.ui.lineEdit.returnPressed.connect(self.changePrompt)
-        
-
-        # """
-        # Brush Tool
-        # """
-        # mainWidgets.brushButton.clicked.connect(self.openBrushMenu)
-
-        # self.BrushMenu = BrushMenuWindow()
-        # self.BrushMenu.ui.brushSizeSlider.valueChanged.connect(self.changeBrushSize)
-
         self.use_brush = False
         
         """
@@ -155,12 +137,7 @@ class ImageFunctions(DNNFunctions):
         
         self.use_erase = False
 
-        """
-        expansion Tool
-        """
-        mainWidgets.jgamButton.clicked.connect(self.checkExpansionTools)
-        self.use_jgam = False
-
+        
         # mainWidgets.classList.itemSelectionChanged.connect(self.convertDNN)
         self.mmseg_status = False
         self.sam_status = False
@@ -190,25 +167,6 @@ class ImageFunctions(DNNFunctions):
 
         self.sam_y_idx = []
         self.sam_x_idx = []
-        
-    def set_button_state(self, use_jgam=False):
-        """
-        Set the state of the buttons
-        """
-        self.use_jgam = use_jgam
-        
-        mainWidgets.jgamButton.setChecked(use_jgam)
-        
-    def checkExpansionTools(self):
-        
-        if hasattr(self, 'imgPath') :
-
-            self.inferenceGroundingDino()
-        
-        else:
-            
-            print(f"no image")     
-
 
     def checkAutoLabelButton(self):
         """
@@ -409,256 +367,6 @@ class ImageFunctions(DNNFunctions):
                 self.set_sam_image()
                 self.sam_x_idx = [] 
                 self.sam_y_idx = []
-
-        ### 
-        # Automatic Joint Gap Measurement 
-        ####
-
-        ## 1. Grounding DINO
-    def inferenceGroundingDino(self, promptVerification=False):
-        
-        self.load_groundingDino(self.groundingDino_config, self.groundingDino_checkpoint)
-        
-        GD_img_source, GD_img = imread_GD(self.imgPath)
-        
-        TEXT_PROMPT = "Steel joint" 
-        BOX_TRESHOLD = 0.25
-        TEXT_TRESHOLD = 0.20
-
-
-        boxes, logits, phrases = self.inference_groundingDino(model=self.groundingDino_model, 
-                                                              image=GD_img,
-                                                              caption=TEXT_PROMPT,
-                                                              box_threshold=BOX_TRESHOLD,
-                                                              text_threshold=TEXT_TRESHOLD,
-                                                              )
-        
-        annotated_frame = annotate_GD(image_source=GD_img_source, boxes=boxes, logits=logits, phrases=phrases)
-        annotated_frame = annotated_frame[...,::-1]
-        index = logits.argmax()
-        box = boxes[index]
-
-        H, W, _ = GD_img_source.shape
-        box_xyxy = box_ops.box_cxcywh_to_xyxy(box) * torch.Tensor([W, H, W, H])
-
-        min_x, min_y, max_x, max_y = box_xyxy.int().tolist()
-        
-        self.GD_min_x = min_x
-        self.GD_min_y = min_y
-        self.GD_max_x = max_x
-        self.GD_max_y = max_y
-
-        if promptVerification:
-            self.promptModel(verification=True)
-        
-        else:
-            self.promptModel()
-        
-        ## 2. Create SAM's Prompt 
-    def promptModel(self, verification=False):
-        
-        if hasattr(self, 'mmseg_model') == False :
-            self.load_mmseg(self.mmseg_config, self.mmseg_checkpoint)
-
-        img = cvtPixmapToArray(self.pixmap)
-        self.GD_img_roi = img[self.GD_min_y:self.GD_max_y, self.GD_min_x:self.GD_max_x, :3]
-        back, joint, gap, logits = self.inference_mmseg(self.GD_img_roi)
-
-        """
-        nomalize the segmentation logits
-        """
-        
-        if len(np.nonzero(self.label[0])) > 0:
-                self.label = np.zeros((img.shape[0], img.shape[1]), dtype=np.uint8)
-                
-        pred_thrs = [0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9] if verification else [self.pred_thr]
-        
-        for thr in pred_thrs:
-
-            prompt_label = np.zeros((img.shape[0], img.shape[1]), dtype=np.uint8)
-
-            # joint
-            joint_logit = logits[1, :, :]
-            joint_score = min_max_normalize(joint_logit)
-            joint_bi = extract_values_above_threshold(joint_score, thr)
-            joint_bi = morphology.remove_small_objects(joint_bi, self.area_thr)
-            joint_bi = morphology.remove_small_holes(joint_bi, self.fill_thr)
-            
-            joint_idx = np.argwhere(joint_bi == 1)
-            joint_y_idx, joint_x_idx = joint_idx[:, 0], joint_idx[:, 1]
-            joint_x_idx = joint_x_idx + self.GD_min_x
-            joint_y_idx = joint_y_idx + self.GD_min_y
-
-            prompt_label[joint_y_idx, joint_x_idx] = 1
-            if thr == 0.8:
-                self.label[joint_y_idx, joint_x_idx] = 1
-            # self.colormap[joint_y_idx, joint_x_idx, :3] = self.label_palette[1]
-
-            # gap
-            gap_logit = logits[2, :, :]
-            gap_score = min_max_normalize(gap_logit)
-            gap_bi = extract_values_above_threshold(gap_score, thr)
-            gap_bi = morphology.remove_small_objects(gap_bi, self.area_thr)
-            gap_bi = morphology.remove_small_holes(gap_bi, self.fill_thr)
-            
-            gap_idx = np.argwhere(gap_bi == 1)
-            gap_y_idx, gap_x_idx = gap_idx[:, 0], gap_idx[:, 1]
-            gap_x_idx = gap_x_idx + self.GD_min_x
-            gap_y_idx = gap_y_idx + self.GD_min_y
-
-            prompt_label[gap_y_idx, gap_x_idx] = 2
-            if thr == 0.8:
-                self.label[gap_y_idx, gap_x_idx] = 2
-            # self.colormap[gap_y_idx, gap_x_idx, :3] = self.label_palette[2]
-
-            # visual
-            img = img[:, :, :3]
-            prompt_colormap = blendImageWithColorMap(img, prompt_label) 
-
-            promptPath = self.imgPath.replace('/leftImg8bit/', '/promptLabelIds/')
-            promptPath = promptPath.replace( '_leftImg8bit.png', f'_prompt({thr})_labelIds.png')
-            promptColormapPath = promptPath.replace(f'_prompt({thr})_labelIds.png', f"_prompt({thr})_color.png")        
-            os.makedirs(os.path.dirname(promptPath), exist_ok=True)
-            
-            imwrite(promptPath, prompt_label) 
-            imwrite_colormap(promptColormapPath, prompt_colormap)
-
-        self.pointSampling(erosion=False)
-
-        ## 2.1. Point Sampling
-    def pointSampling(self, erosion=False):
-        
-        joint = self.label == 1
-        gap = self.label == 2
-        
-        if erosion == True:
-            joint = morphology.erosion(joint, morphology.square(15))
-            gap = morphology.erosion(gap, morphology.square(15))
-                
-        # img = np.array(Image.open(self.imgPath))
-
-        # joint_col_coord = column_based_sampling(img, joint, num_samples=10, num_columns=10)
-        # joint_row_coord = width_based_sampling(img, joint, num_samples=10, num_rows=10)
-
-        # joint_coord = np.concatenate((joint_col_coord, joint_row_coord), axis=0)
-        # joint_coord = joint_coord[:, [1,0]]
-        # joint_label = np.zeros((joint_coord.shape[0]), dtype=int)
-
-        # gap_col_coord = column_based_sampling(img, gap, num_samples=10, num_columns=10)
-        # gap_row_coord = width_based_sampling(img, gap, num_samples=10, num_rows=10)
-
-        # gap_coord = np.concatenate((gap_col_coord, gap_row_coord), axis=0)
-        # gap_coord = gap_coord[:, [1,0]]
-        # gap_label = np.ones((gap_coord.shape[0]), dtype=int)
-
-        
-        
-        # top6 Area centroid point sampling
-        
-        # joint
-        top6_joint = getTop6Skeletonize(joint, onlycenter=True)
-        self.top6_joint = np.array(top6_joint)
-        self.top6_joint_label = np.zeros((self.top6_joint.shape[0]), dtype=int)
-        # gap
-        top6_gap = getTop6Skeletonize(gap, onlycenter=True)
-        self.top6_gap = np.array(top6_gap)
-        self.top6_gap_label = np.ones((self.top6_gap.shape[0]), dtype=int)
-
-        input_point = np.concatenate((self.top6_gap, self.top6_joint), axis=0)
-        input_label = np.concatenate((self.top6_gap_label, self.top6_joint_label), axis=0)
-        
-        input_box = np.array([self.GD_min_x, self.GD_min_y, self.GD_max_x, self.GD_max_y])
-
-        self.inferenceSAM(input_point, input_label, input_box)
-    
-        ## 3. SAM inference
-    def inferenceSAM(self, input_point, input_label, input_box):
-        
-        if hasattr(self, 'sam_model') == False :
-            self.load_sam(self.sam_checkpoint) 
-
-        img = cvtPixmapToArray(self.pixmap)
-        img = img[:, :, :3]
-                
-        self.sam_predictor.set_image(img)
-        
-        
-        masks, scores, logits = self.sam_predictor.predict(
-            point_coords=input_point,
-            point_labels=input_label,
-            box=input_box, 
-            multimask_output=True,
-        )
-
-        mask = masks[np.argmax(scores), :, :]
-        self.sam_mask_input = logits[np.argmax(scores), :, :]
-
-        # update label with result
-        idx = np.argwhere(mask == 1)
-        y_idx, x_idx = idx[:, 0], idx[:, 1]
-
-        self.GD_sam_y_idx = y_idx
-        self.GD_sam_x_idx = x_idx
-        
-        self.label[self.label==1] = 0
-        
-        self.updateColorMap()
-
-        self.label[y_idx, x_idx] = 2
-        self.colormap[y_idx, x_idx, :3] = self.label_palette[2]
-
-        imwrite(self.labelPath, self.label)
-
-        _colormap = copy.deepcopy(self.colormap)
-
-        for joint in self.top6_joint:
-            
-            # cv2.circle(_colormap, (joint[0], joint[1]), 50, (0, 0, 255, 255), 9)
-            cv2.circle(_colormap, (joint[0], joint[1]), 9, (0, 0, 255, 255), -1)
-        for gap in self.top6_gap:
-            
-            # cv2.circle(_colormap, (gap[0], gap[1]), 50, (255, 0, 0, 255), 9)
-            cv2.circle(_colormap, (gap[0], gap[1]), 9, (255, 0, 0, 255), -1)
-        
-
-
-        self.color_pixmap = QPixmap(cvtArrayToQImage(_colormap))
-        self.color_pixmap_item.setPixmap(QPixmap())
-        self.color_pixmap_item.setPixmap(self.color_pixmap)
-
-        sam_colormap = blendImageWithColorMap(img, self.label) 
-        img = imread(self.imgPath)
-
-        for joint in self.top6_joint:
-            
-            cv2.circle(sam_colormap, (joint[0], joint[1]), 9, (255, 0, 0, 255), -1)
-            cv2.circle(img, (joint[0], joint[1]), 9, (255, 0, 0, 255), -1)
-        for gap in self.top6_gap:
-            
-            cv2.circle(sam_colormap, (gap[0], gap[1]), 9, (0, 0, 255, 255), -1)
-            cv2.circle(img, (gap[0], gap[1]), 9, (0, 0, 255, 255), -1)
-        
-        
-        colormapPath = os.path.dirname(self.labelPath)
-        colormapName = os.path.basename(self.labelPath)
-        colormapPath = os.path.dirname(colormapPath)
-        colormapPath = os.path.dirname(colormapPath)
-        colormapPath = os.path.join(colormapPath, "JGAM_colormap")
-        os.makedirs(colormapPath, exist_ok=True)
-        colormapPath = os.path.join(colormapPath, colormapName)
-
-        pointName = colormapName.replace("_labelIds.png", "point.png")
-        pointmapPath = os.path.join(os.path.dirname(colormapPath), pointName)
-
-
-        imwrite_colormap(colormapPath, sam_colormap)
-        cv2.imwrite(pointmapPath, img)
-        
-        print(f"colormapPath: {colormapPath}, pointmapPath:{pointmapPath}")
-        
-
-
-
         
     def drawRectangle(self, event):
